Continuing integration of euinstance, removed some debug information, and service manager

+++ eutester/__init__.py
@@ -51,15 +51,17 @@ from threading import Thread
 from boto.ec2.regioninfo import RegionInfo
 from boto.s3.connection import OrdinaryCallingFormat
 
-from bm_machine import bm_machine
+from machine import machine
 import eulogger
+from euservice import EuserviceManager
+
 
 class TimeoutFunctionException(Exception): 
     """Exception to raise on a timeout""" 
     pass 
 
 class Eutester(object):
-    def __init__(self, config_file=None, hostname=None, password=None, keypath=None, credpath=None, aws_access_key_id=None, aws_secret_access_key = None, account="eucalyptus",  user="admin", boto_debug=0):
+    def __init__(self, config_file=None, password=None, keypath=None, credpath=None, aws_access_key_id=None, aws_secret_access_key = None, account="eucalyptus",  user="admin", boto_debug=0):
         """  
         EUCADIR => $eucadir, 
         VERIFY_LEVEL => $verify_level, 
@@ -82,7 +84,7 @@ class Eutester(object):
         self.current_ssh = "clc"
         self.boto_debug = boto_debug
         self.ssh = None
-        self.hostname = hostname
+        self.sftp = None
         self.password = password
         self.keypath = keypath
         self.credpath = credpath
@@ -104,7 +106,7 @@ class Eutester(object):
         self.logging_thread = False
         
         ### Eutester logs
-        self.logger = eulogger.Eulogger(name= "eutester")
+        self.logger = eulogger.Eulogger(identifier="localhost")
         self.debug = self.logger.log.debug
         self.critical = self.logger.log.critical
         self.info = self.logger.log.info
@@ -118,6 +120,8 @@ class Eutester(object):
         self.cc_log_channel= None
         self.nc_log_channel= None
         
+        self.clc_index = 0
+
         ### If I have a config file
         ### PRIVATE CLOUD
         if self.config_file != None:
@@ -127,35 +131,30 @@ class Eutester(object):
             ### Set the eucapath
             if "REPO" in self.config["machines"][0].source:
                 self.eucapath="/"
-            ### swap in the hostname of the component 
-            self.hostname = self.swap_component_hostname(self.current_ssh)
-            self.debug("Hostname for SSH connection: " + self.hostname)
-            self.hypervisor = self.get_hypervisor()
-            
-        ## IF I WASNT PROVIDED KEY TRY TO GET THEM FROM THE EUCARC IN CREDPATH
-        ### PRIVATE CLOUD
-        if (self.password != None) or (self.keypath != None):
-            client = paramiko.SSHClient()
-            client.set_missing_host_key_policy(paramiko.AutoAddPolicy())
-            self.debug("Issuing SSH connection root@" +  self.hostname )            
-            if keypath == None:   
-                client.connect(self.hostname, username="root", password=password, timeout= self.timeout)
-            else:
-                client.connect(self.hostname,  username="root", key_filename=keypath, timeout= self.timeout)
-            self.ssh = client
-            self.sftp = self.ssh.open_sftp()
-        
-        ### If i have an ssh session and its to the clc
-        ### Private cloud with root access
-        if (self.credpath == None) and (self.ssh != None) and (self.password != None):
-                self.credpath = self.get_credentials(account,user) 
-        
+            #self.hypervisor = self.get_hypervisor()
+            ### No credpath but does have password and an ssh connection to the CLC
+            ### Private cloud with root access 
+            ### Need to get credentials for the user if there arent any passed in
+            ### Need to create service manager for user if we have an ssh connection and password
+            if (self.password != None):
+                self.clc = self.get_component_machines("clc")[self.clc_index]
+                if self.credpath is None:
+                    ### TRY TO GET CREDS ON FIRST CLC if it fails try on second listed clc, if that fails weve hit a terminal condition
+                    try:
+                        self.sftp = self.clc.ssh.connection.open_sftp()
+                        self.credpath = self.get_credentials(account,user)
+                    except Exception, e:
+                        self.swap_clc()
+                        self.sftp = self.clc.ssh.connection.open_sftp()
+                        self.credpath = self.get_credentials(account,user)
+                self.service_manager = EuserviceManager(self)
+                self.clc = self.service_manager.get_enabled_clc().machine
 
-        ### If i have a credpath
+        ### Pull the access and secret keys from the eucarc
         if (self.credpath != None):         
             aws_access_key_id = self.get_access_key()
             aws_secret_access_key = self.get_secret_key()
-                 
+                    
         ### If you have credentials for the boto connections, create them
         if (aws_access_key_id != None) and (aws_secret_access_key != None):
            self.ec2 = boto.connect_ec2(aws_access_key_id=aws_access_key_id,
@@ -184,6 +183,14 @@ class Eutester(object):
     
     def __del__(self):
         self.logging_thread = False
+        
+    def swap_clc(self):
+        if self.clc_index is 0:
+            self.clc = self.get_component_machines("clc")[1]
+            self.clc_index = 1
+        else:
+            self.clc = self.get_component_machines("clc")[0]
+            self.clc_index = 0
     
     def read_config(self, filepath):
         """ Parses the config file at filepath returns a dictionary with the config
@@ -235,9 +242,16 @@ class Eutester(object):
                 machine_dict["source"] = machine_details[4]
                 machine_dict["components"] = map(str.lower, machine_details[5].strip('[]').split())
                 ### ADD the machine to the array of machine
-                machine = bm_machine(machine_dict["hostname"], machine_dict["distro"], machine_dict["distro_ver"], machine_dict["arch"], machine_dict["source"], machine_dict["components"])
-                machines.append(machine)
-               # print machine
+                cloud_machine = machine(   machine_dict["hostname"], 
+                                        machine_dict["distro"], 
+                                        machine_dict["distro_ver"], 
+                                        machine_dict["arch"], 
+                                        machine_dict["source"], 
+                                        machine_dict["components"],
+                                        self.password,
+                                        self.keypath
+                                        )
+                machines.append(cloud_machine)
             if re.search("network",line, re.IGNORECASE):
                 config_hash["network"] = line.split()[1].lower()
         config_hash["machines"] = machines 
@@ -274,6 +288,14 @@ class Eutester(object):
             raise Exception("Could not find component "  + component + " in list of machines")
         else:
              return machines_with_role[0]
+    
+    def get_machine_by_ip(self, hostname):
+         machines = [machine for machine in self.config['machines'] if re.search(hostname, machine.hostname)]
+         if len(machines) == 0:
+            self.fail("Could not find machine at "  + hostname + " in list of machines")
+            return None
+         else:
+             return machines[0]
          
     def get_component_machines(self, component):
         #loop through machines looking for this component type
@@ -297,14 +319,24 @@ class Eutester(object):
            Defaults to admin@eucalyptus 
         """
         admin_cred_dir = "eucarc-" + account + "-" + user
+        
+        ### SETUP directory remotely
         self.sys("rm -rf " + admin_cred_dir)
         self.sys("mkdir " + admin_cred_dir)
+        
+        ### Download credentials from Active CLC
         cmd_download_creds = self.eucapath + "/usr/sbin/euca_conf --get-credentials " + admin_cred_dir + "/creds.zip " + "--cred-user "+ user +" --cred-account " + account 
-        cmd_setup_cred_dir = ["rm -rf " + admin_cred_dir,"mkdir " + admin_cred_dir ,  cmd_download_creds, "unzip -o " + admin_cred_dir + "/creds.zip " + "-d " + admin_cred_dir]
-        for cmd in cmd_setup_cred_dir:         
-            stdout = self.sys(cmd, verbose=1)
+
+        if self.found( cmd_download_creds, ""):
+            raise IOError("Error downloading credentials")
+        if self.found( "unzip -o " + admin_cred_dir + "/creds.zip " + "-d " + admin_cred_dir, "cannot find zipfile directory"):
+            raise IOError("Empty ZIP file returned by CLC")
+        
+        ### SETUP directory locally
         os.system("rm -rf " + admin_cred_dir)
         os.mkdir(admin_cred_dir)
+        
+        ### DOWNLOAD creds from clc
         self.sftp.get(admin_cred_dir + "/creds.zip" , admin_cred_dir + "/creds.zip")
         os.system("unzip -o " + admin_cred_dir + "/creds.zip -d " + admin_cred_dir )
         return admin_cred_dir
@@ -430,49 +462,13 @@ class Eutester(object):
     def handle_timeout(self, signum, frame): 
         raise TimeoutFunctionException()
     
-    def sys(self, cmd, verbose=1, timeout=-2):
+    def sys(self, cmd, verbose=True):
         """ By default will run a command on the CLC machine, the connection used can be changed by passing a different hostname into the constructor
             For example:
             instance = Eutester( hostname=instance.ip_address, keypath="my_key.pem")
             instance.sys("mount") # check mount points on instance and return the output as a list
         """
-        cmd = str(cmd)
-        # default timeout is to use module-defined timeout
-        # -1 should be reserved for "no timeout" option
-        if timeout == -2:
-            timeout = self.timeout
-        time.sleep(self.delay)
-        old = signal.signal(signal.SIGALRM, self.handle_timeout) 
-        signal.alarm(timeout) 
-        cur_time = time.strftime("%I:%M:%S", time.gmtime())
-        output = []
-        std_out_return = []
-        if verbose:
-            if self.ssh == None:
-                self.hostname ="localhost"
-            self.debug( "[root@" + str(self.hostname) + "]# " + cmd)
-        try:
-            
-            if self.ssh == None:
-                std_out_return = self.local(cmd)
-            else:
-                stdin_ls, stdout_ls, stderr_ls = self.ssh.exec_command("ls")
-                ls_result = stdout_ls.readlines()
-                if self.credpath != None:
-                    for item in ls_result: 
-                        if re.match(self.credpath,item): 
-                            cmd = ". " + self.credpath + "/eucarc && " + cmd
-                            break
-                stdin, stdout, stderr = self.ssh.exec_command(cmd)
-                std_out_return = stdout.readlines() 
-                output = std_out_return
-        except Exception, e: 
-            self.fail("Command timeout after " + str(timeout) + " seconds\nException:" + str(e)) 
-            return []
-        signal.alarm(0)      
-        if verbose:
-            self.debug("".join(std_out_return))
-        return std_out_return
+        return self.clc.sys(cmd, verbose=verbose)
 
     def local(self, cmd):
         """ Run a command locally on the tester"""
@@ -485,7 +481,7 @@ class Eutester(object):
     def found(self, command, regex, local=False):
         """ Returns a Boolean of whether the result of the command contains the regex"""
         if local:
-            result = self.local(cmd)
+            result = self.local(command)
         else:
             result = self.sys(command)
         for line in result:
@@ -504,7 +500,6 @@ class Eutester(object):
         return list(set(list1)-set(list2))
     
     def fail(self, message):
-        #self.debug( "[TEST_REPORT] FAILED: " + message)
         self.critical("[TEST_REPORT] FAILED: " + message)
         self.fail_log.append(message)
         self.fail_count += 1
@@ -545,7 +540,6 @@ class Eutester(object):
         s  = "+++++++++++++++++++++++++++++++++++++++++++++++++++++\n"
         s += "+" + "Eucateser Configuration" + "\n"
         s += "+" + "+++++++++++++++++++++++++++++++++++++++++++++++\n"
-        s += "+" + "Host:" + self.hostname + "\n"
         s += "+" + "Config File: " + self.config_file +"\n"
         s += "+" + "Fail Count: " +  str(self.fail_count) +"\n"
         s += "+" + "Eucalyptus Path: " +  str(self.eucapath) +"\n"
